email_agent/requirements.txt
google-adk>=0.1.0
python-dotenv>=1.0.0
mcp-email-server>=0.1.0
